+++ src/prefect/environments/storage/__init__.py
@@ -24,6 +24,7 @@ from prefect.environments.storage.bytes import Bytes
 from prefect.environments.storage.local import Local
 from prefect.environments.storage.memory import Memory
 from prefect.environments.storage.gcs import GCS
+from prefect.environments.storage.s3 import S3
 
 
 def get_default_storage_class() -> type: