+++ src/prefect/serialization/storage.py
@@ -4,7 +4,7 @@ import marshmallow
 from marshmallow import fields, post_load
 
 import prefect
-from prefect.environments.storage import Bytes, Docker, GCS, Local, Memory, Storage
+from prefect.environments.storage import Bytes, Docker, GCS, Local, Memory, Storage, S3
 from prefect.utilities.serialization import Bytes as BytesField
 from prefect.utilities.serialization import ObjectSchema, OneOfSchema
 
@@ -78,6 +78,22 @@ class LocalSchema(ObjectSchema):
         return base_obj
 
 
+class S3Schema(ObjectSchema):
+    class Meta:
+        object_class = S3
+
+    bucket = fields.String(allow_none=False)
+    key = fields.String(allow_none=True)
+    flows = fields.Dict(key=fields.Str(), values=fields.Str())
+
+    @post_load
+    def create_object(self, data: dict, **kwargs: Any) -> S3:
+        flows = data.pop("flows", dict())
+        base_obj = super().create_object(data)
+        base_obj.flows = flows
+        return base_obj
+
+
 class MemorySchema(ObjectSchema):
     class Meta:
         object_class = Memory
@@ -96,4 +112,5 @@ class StorageSchema(OneOfSchema):
         "Local": LocalSchema,
         "Memory": MemorySchema,
         "Storage": BaseStorageSchema,
+        "S3": S3Schema,
     }